Adicionado sistema de niveis

+++ src/br/com/dio/desafio/dominio/Conteudo.java
@@ -2,7 +2,7 @@ package br.com.dio.desafio.dominio;
 
 public abstract class Conteudo {
 
-    protected static final double XP_PADRAO = 10d;
+    protected static final int XP_PADRAO = 10;
     private String titulo;
     private String descricao;
